chatsystemapp/requirements.txt
Django==6.0
asgiref==3.11.0
sqlparse==0.5.4
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1
redis==7.1.0
